move main code to testing_utils, improve the set_verbosity_all() function
