+++ tests/conftest.py
@@ -13,17 +13,12 @@ git_repo_path = abspath(join(dirname(dirname(__file__)), "src"))
 sys.path.insert(1, git_repo_path)
 
 # import local modules after fixing up sys.path
-if 1:  # flake be quiet
-    from transformers.utils.logging import (
-        logging_level_str_to_code,
-        logging_levels_as_strings,
-        set_global_logging_level,
-    )
+from transformers.testing_utils import logging_level_str_to_code, logging_levels_as_strings, set_verbosity_all  # NOQA
 
 
 def pytest_addoption(parser):
     parser.addoption(
-        "--loglevel",
+        "--log-level-all",
         type=str,
         default=False,
         choices=logging_levels_as_strings(),
@@ -33,7 +28,9 @@ def pytest_addoption(parser):
 
 @pytest.fixture(scope="session", autouse=True)
 def run_this_before_each_test(request):
-    # set the loglevel for all loggers to the desired level
-    loglevel = request.config.getoption("--loglevel")
+    # set the loglevel for all loggers (not just transformers) to the desired level
+    # note: as --log-level-all=error override is intended to be run only for one test
+    # under debugging, this practically will have no impact on the test suite at large
+    loglevel = request.config.getoption("--log-level-all")
     if loglevel:
-        set_global_logging_level(level=logging_level_str_to_code(loglevel))
+        set_verbosity_all(level=logging_level_str_to_code(loglevel))
